+++ src/borders.h
@@ -1,8 +1,7 @@
 #ifndef _BORDERS_H_
 #define _BORDERS_H_
 
-#include "types.h"
-#include "f3ddkr.h"
+#include "PR/gbi.h"
 
 void divider_draw(Gfx **dList);
 void divider_clear_coverage(Gfx **dList);
