+++ src/game_text.c
@@ -6,6 +6,8 @@
 #include "game.h"
 #include "joypad.h"
 #include "PRinternal/viint.h"
+#include "asset_loading.h"
+#include "thread3_main.h"
 
 /************ .data ************/
