+++ src/joypad.c
@@ -1,5 +1,7 @@
 #include "joypad.h"
 #include "game.h"
+#include "menu.h"
+#include "thread3_main.h"
 
 s32 sNoControllerPluggedIn =
     FALSE; // Looks to be a boolean for whether a controller is plugged in. FALSE if plugged in, and TRUE if not.
