+++ src/object_functions.c
@@ -30,6 +30,7 @@
 #include "printf.h"
 #include "PRinternal/viint.h"
 #include "common.h"
+#include "thread3_main.h"
 
 /************ .data ************/
