+++ src/vehicle_tricky.c
@@ -11,6 +11,7 @@
 #include "game.h"
 #include "particles.h"
 #include "math_util.h"
+#include "thread3_main.h"
 
 /************ .data ************/
